fix: Fix OBJ_Decoration's constructor to accept double col and row

+++ src/object/OBJ_Decoration.java
@@ -7,14 +7,14 @@ import main.GamePanel;
 
 public class OBJ_Decoration extends Entity {
 	
-	public OBJ_Decoration(GamePanel gp, String imagePath, int width, int height, int col, int row) {
+	public OBJ_Decoration(GamePanel gp, String imagePath, int width, int height, double col, double row) {
 		super(gp);
 		down1 = setup(imagePath, width, height);
 		
 		solidArea = new Rectangle (0, 0, width, height);
 		collision = true;
 		
-		this.worldX = gp.tileSize * col;
-		this.worldY = gp.tileSize * row;
+		this.worldX = (int) (gp.tileSize * col);
+		this.worldY = (int) (gp.tileSize * row);
 	}
 }
